Create parent component
