+++ src/Sylius/Bundle/ShopBundle/templates/product/show/page/info/summary.html.twig
@@ -1,5 +1,5 @@
-<div class="col-12 col-lg-4 order-lg-1">
+<div class="col-12 col-lg-4 order-lg-1" {{ attributes }}>
     <div class="sticky-top pt-2" style="top: 3rem;" {{ sylius_test_html_attribute('product-box') }}>
-        {% hook 'summary' %}
+        {% hook 'summary' with { product, variant } %}
     </div>
 </div>
