+++ src/Sylius/Bundle/ShopBundle/templates/product/show/page/info/summary/price/lowest_price_before_discount.html.twig
@@ -1,6 +1,6 @@
 {% import "@SyliusShop/shared/macro/money.html.twig" as money %}
 
-{% set variant = hookable_metadata.context.product|sylius_resolve_variant %}
+{% set variant = hookable_metadata.context.variant %}
 {% set has_discount = hookable_metadata.context.has_discount %}
 {% set days = hookable_metadata.context.days %}
 {% set has_lowest_price = hookable_metadata.context.has_lowest_price %}
