+++ src/Sylius/Bundle/ShopBundle/templates/product/show/page/info/summary/variant_selection/add_to_cart/variants/table/body/selection.html.twig
@@ -1,3 +1,3 @@
 <td>
-    {{ form_widget(hookable_metadata.context.form.cartItem.variant[hookable_metadata.context.key], {'label': false}) }}
+    {{ form_widget(hookable_metadata.context.form.cartItem.variant[hookable_metadata.context.key], {'label': false, 'attr': { 'data-action': 'live#action', 'data-live-action-param': 'variantChanged' } }) }}
 </td>
